Refactored mixins
DRYed up a bunch of code
