+++ addon/mixins/controllers/model-records/columns.js
@@ -1,0 +1,9 @@
+import Ember from 'ember';
+
+export default Ember.Mixin.create({
+  needs: ['model-records'],
+  columns: Ember.computed.mapBy('controllers.model-records.columns', 'name'),
+  filteredColumns: Ember.computed.filter('columns', function(name) {
+    return name !== 'id';
+  })
+});
