+++ addon/mixins/routes/model-records/edit.js
@@ -1,28 +1,12 @@
 import Ember from 'ember';
+import WriteMixin from 'ember-admin/mixins/routes/model-records/write';
 
-export default Ember.Mixin.create({
+export default Ember.Mixin.create(WriteMixin, {
   model: function(params) {
     return this.admin.store.find(this.modelFor('model-records').name, params.id);
   },
-  renderTemplate: function() {
-    var templatePath = 'admin/edit/' + this.modelFor('model-records').name;
-
-    if (this.container.resolve('template:'+templatePath)) {
-      this.render(templatePath);
-    } else {
-      this.render();
-    }
-  },
+  templateAdminPath: 'admin/edit',
   actions: {
-    save: function(callback) {
-      var _this = this;
-      var promise = this.get('controller.model').save();
-      callback(promise);
-
-      promise.then(function() {
-        _this.transitionTo('model-records', _this.modelFor('model-records').name);
-      });
-    },
     destroyRecord: function(callback) {
       var _this = this;
       var canDestroy = window.confirm("Are you sure you want to destory this record?");
@@ -42,11 +26,5 @@ export default Ember.Mixin.create({
         callback(promise);
       }
     },
-    cancel: function() {
-      this.transitionTo('model-records', this.modelFor('model-records').name);
-    }
-  },
-  willTransitionConfirm: function() {
-    return window.confirm("You have unsaved changes. Are you sure you want to continue?");
   }
 });
